Adds Animator export documentation
Adds documentation for the Animator export process in the migration guide.
This export is mandatory for projects with animated sprites, detailing
the steps required and consequences of skipping this step.
Includes images to visually guide the user through the process.
Increments the table of contents as a result of adding the Animator Export documentation.